Update the heuristic evaluation Manager page styling to match the user testing Manager page
